test: fold web provider test suites into one file
